BantayAyuda/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.5.0
requests==2.32.3
catboost==1.2.2
google-generativeai==0.3.2
pandas==2.2.0
numpy==1.26.3
django-leaflet==0.28.3
psycopg2-binary==2.9.9

